Move `httpStatusCodeFor` into type
