+++ packages/framework-provider-aws/src/library/api-gateway-io.ts
@@ -1,22 +1,5 @@
 import { APIGatewayProxyResult } from 'aws-lambda'
-import {
-  InvalidParameterError,
-  NotAuthorizedError,
-  NotFoundError,
-  InvalidVersionError,
-  toClassTitle,
-} from '@boostercloud/framework-types'
-
-function httpStatusCodeFor(error: Error): number {
-  const errorToHTTPCode: Record<string, number> = {
-    [InvalidParameterError.name]: 400,
-    [NotAuthorizedError.name]: 401,
-    [NotFoundError.name]: 404,
-    [InvalidVersionError.name]: 422,
-  }
-
-  return errorToHTTPCode[error.constructor.name] ?? 500
-}
+import { httpStatusCodeFor, toClassTitle } from '@boostercloud/framework-types'
 
 export async function requestFailed(error: Error): Promise<APIGatewayProxyResult> {
   const statusCode = httpStatusCodeFor(error)
